requirements.txt
python-dotenv
boto3
websockets
pyarrow
pyspark
confluent-kafka
duckdb
pandas
apache-airflow-providers-apache-spark
apache-airflow-providers-common-sql